Merged PR 172: Give both device and patient information for all windows. And enforce windows not sharing two devices or two patients.
This branch introduces a "device_patient_tuples" source type to the validated sources data structure for the iterator.

It involves reading the device_patient table and looking for overlaps between it and the requested data from the definition.

It also now brings that raw interval information into each batch request ensuring that get_data is only called within the regions that are mapped one-to-one with devices and patients. Areas where we have data but it would belong to a different patient within a window are replaced with NaN values.